Fix wheelEvent parameter type

+++ source/app/ui/graphcommoninteractor.cpp
@@ -438,7 +438,7 @@ NodeId GraphCommonInteractor::nodeIdNearPosition(const QPoint& localPosition) co
                                            frustum, ray.origin());
 }
 
-void GraphCommonInteractor::wheelEvent(const QPoint& pos, float angle)
+void GraphCommonInteractor::wheelEvent(const QPoint& pos, int angle)
 {
     if(_graphRenderer->transition().active())
         return;
